Tasks/requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.28
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.0
Flask-Bcrypt==1.0.1
werkzeug==2.2.2 